Include creation time in benchmark comparison for Completable and Listenable Future
This is because there is no clear demarcation between create and execution time. Previously, a lot
of tasks were already executing/completed by the time the clock was started, leading to lower
values. With this change, the values are slightly higher because we also include the create time.
The actual performance for these two futures should be somewhere in between.